+++ examples/basic/src/model-provider/openai/openai-generate-structure-chat-example.ts
@@ -1,22 +1,23 @@
+import dotenv from "dotenv";
 import {
   OpenAIChatFunctionPrompt,
   OpenAIChatMessage,
   OpenAIChatModel,
-  generateJson,
+  ZodStructureDefinition,
+  generateStructure,
 } from "modelfusion";
-import dotenv from "dotenv";
 import { z } from "zod";
 
 dotenv.config();
 
 async function main() {
-  const sentiment = await generateJson(
+  const sentiment = await generateStructure(
     new OpenAIChatModel({
       model: "gpt-3.5-turbo",
       temperature: 0,
       maxCompletionTokens: 50,
     }),
-    {
+    new ZodStructureDefinition({
       name: "sentiment" as const,
       description: "Write the sentiment analysis",
       schema: z.object({
@@ -24,8 +25,8 @@ async function main() {
           .enum(["positive", "neutral", "negative"])
           .describe("Sentiment."),
       }),
-    },
-    OpenAIChatFunctionPrompt.forSchemaCurried([
+    }),
+    OpenAIChatFunctionPrompt.forStructureCurried([
       OpenAIChatMessage.system(
         "You are a sentiment evaluator. " +
           "Analyze the sentiment of the following product review:"