+++ examples/basic/src/model-provider/openai/openai-generate-structure-or-text-chat-example.ts
@@ -1,10 +1,11 @@
+import dotenv from "dotenv";
 import {
   OpenAIChatFunctionPrompt,
   OpenAIChatMessage,
   OpenAIChatModel,
-  generateJsonOrText,
+  ZodStructureDefinition,
+  generateStructureOrText,
 } from "modelfusion";
-import dotenv from "dotenv";
 import { z } from "zod";
 
 dotenv.config();
@@ -14,10 +15,10 @@ async function main() {
   // const query = "Where does Kevin work?";
   // const query = "Tell me something random.";
 
-  const { schema, value, text } = await generateJsonOrText(
+  const { structure, value, text } = await generateStructureOrText(
     new OpenAIChatModel({ model: "gpt-3.5-turbo", maxCompletionTokens: 1000 }),
     [
-      {
+      new ZodStructureDefinition({
         name: "getCurrentWeather" as const, // mark 'as const' for type inference
         description: "Get the current weather in a given location",
         schema: z.object({
@@ -26,19 +27,21 @@ async function main() {
             .describe("The city and state, e.g. San Francisco, CA"),
           unit: z.enum(["celsius", "fahrenheit"]).optional(),
         }),
-      },
-      {
-        name: "getContactInformation" as const,
+      }),
+      new ZodStructureDefinition({
+        name: "getContactInformation" as const, // mark 'as const' for type inference
         description: "Get the contact information for a given person",
         schema: z.object({
           name: z.string().describe("The name of the person"),
         }),
-      },
+      }),
     ],
-    OpenAIChatFunctionPrompt.forSchemasCurried([OpenAIChatMessage.user(query)])
+    OpenAIChatFunctionPrompt.forStructuresCurried([
+      OpenAIChatMessage.user(query),
+    ])
   );
 
-  switch (schema) {
+  switch (structure) {
     case "getCurrentWeather": {
       const { location, unit } = value;
       console.log("getCurrentWeather", location, unit);