+++ examples/basic/src/util/schema/generate-structure-unchecked-json-schema-example.ts
@@ -1,31 +1,39 @@
+import dotenv from "dotenv";
 import {
   OpenAIChatFunctionPrompt,
   OpenAIChatMessage,
   OpenAIChatModel,
-  generateJson,
+  UncheckedJsonSchemaStructureDefinition,
+  generateStructure,
 } from "modelfusion";
-import dotenv from "dotenv";
-import { z } from "zod";
 
 dotenv.config();
 
 async function main() {
-  const sentiment = await generateJson(
+  const sentiment = await generateStructure(
     new OpenAIChatModel({
       model: "gpt-3.5-turbo",
       temperature: 0,
       maxCompletionTokens: 50,
     }),
-    {
-      name: "sentiment" as const,
+    new UncheckedJsonSchemaStructureDefinition({
+      name: "sentiment",
       description: "Write the sentiment analysis",
-      schema: z.object({
-        sentiment: z
-          .enum(["positive", "neutral", "negative"])
-          .describe("Sentiment."),
-      }),
-    },
-    OpenAIChatFunctionPrompt.forSchemaCurried([
+      jsonSchema: {
+        $schema: "http://json-schema.org/draft-07/schema#",
+        type: "object",
+        properties: {
+          sentiment: {
+            type: "string",
+            enum: ["positive", "neutral", "negative"],
+            description: "Sentiment.",
+          },
+        },
+        required: ["sentiment"],
+        additionalProperties: false,
+      },
+    }),
+    OpenAIChatFunctionPrompt.forStructureCurried([
       OpenAIChatMessage.system(
         "You are a sentiment evaluator. " +
           "Analyze the sentiment of the following product review:"