+++ src/core/index.ts
@@ -8,3 +8,4 @@ export * from "./GlobalFunctionObservers.js";
 export * from "./Run.js";
 export * from "./Vector.js";
 export * from "./api/index.js";
+export * from "./structure/index.js";