+++ src/core/structure/StructureDefinition.ts
@@ -1,0 +1,7 @@
+import { Schema } from "./Schema.js";
+
+export interface StructureDefinition<NAME extends string, STRUCTURE> {
+  name: NAME;
+  description?: string;
+  schema: Schema<STRUCTURE>;
+}