+++ src/core/structure/UncheckedJsonSchemaSchema.ts
@@ -1,0 +1,17 @@
+import { Schema } from "./Schema.js";
+
+export class UncheckedJsonSchemaSchema<STRUCTURE> implements Schema<STRUCTURE> {
+  constructor(private readonly jsonSchema: unknown) {}
+
+  validate(
+    value: unknown
+  ): { success: true; value: STRUCTURE } | { success: false; error: unknown } {
+    return { success: true, value: value as STRUCTURE };
+  }
+
+  getJsonSchema(): unknown {
+    return this.jsonSchema;
+  }
+
+  readonly _type: STRUCTURE;
+}