+++ src/model-function/generate-json/SchemaDefinition.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export type SchemaDefinition<NAME extends string, STRUCTURE> = {
-  name: NAME;
-  description?: string;
-  schema: z.Schema<STRUCTURE>;
-};